Move TitleText into the Text components folder
